+++ packages/webapp/core/README.md
@@ -1,0 +1,3 @@
+# Webapp Core
+
+This package contains the core webapp
