+++ packages/webapp/core/package.json
@@ -1,0 +1,19 @@
+{
+	"name": "core",
+	"scripts": {
+		"build": "rollup -c",
+		"dev": "rollup -c -w"
+	},
+	"devDependencies": {
+		"@rollup/plugin-commonjs": "^14.0.0",
+		"@rollup/plugin-node-resolve": "^8.4.0",
+		"@rollup/plugin-sucrase": "^3.1.0",
+		"@rollup/plugin-typescript": "^5.0.2",
+		"prettier": "^2.0.5",
+		"rollup": "^2.23.1",
+		"rollup-plugin-livereload": "^1.3.0",
+		"rollup-plugin-terser": "^6.1.0",
+		"tslib": "^2.0.1",
+		"typescript": "^3.9.7"
+	}
+}
