+++ packages/webapp/core/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../../tsconfig.base.json",
+	"compilerOptions": {
+		"rootDir": "source",
+		"outDir": "build",
+		"sourceMap": true
+	}
+}
