+++ packages/webapp/deps/README.md
@@ -1,0 +1,3 @@
+# Webapp Dependencies
+
+We package all our dependencies into a single bundle here
