+++ packages/webapp/deps/package.json
@@ -1,0 +1,26 @@
+{
+	"name": "deps",
+	"main": "build/bundle.prod.js",
+	"types": "build/index.d.ts",
+	"scripts": {
+		"build": "rollup -c"
+	},
+	"dependencies": {
+		"pixi.js": "^5.3.3",
+		"react": "^16.13.1",
+		"react-dom": "^16.13.1",
+		"regexparam": "^1.3.0"
+	},
+	"devDependencies": {
+		"@rollup/plugin-commonjs": "^14.0.0",
+		"@rollup/plugin-node-resolve": "^8.4.0",
+		"@rollup/plugin-typescript": "^5.0.2",
+		"@types/react": "^16.9.46",
+		"@types/react-dom": "^16.9.8",
+		"prettier": "^2.0.5",
+		"rollup": "^2.23.1",
+		"rollup-plugin-terser": "^6.1.0",
+		"tslib": "^2.0.1",
+		"typescript": "^3.9.7"
+	}
+}
