+++ packages/webapp/deps/source/index.ts
@@ -1,0 +1,5 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import * as PixiJS from 'pixi.js'
+
+export { React, ReactDOM, PixiJS }
