--- packages/webapp/jest.config.js
@@ -1,15 +1,0 @@
-module.exports = {
-	clearMocks: true,
-
-	coverageDirectory: '../coverage',
-
-	coverageReporters: ['json', 'text'],
-
-	moduleFileExtensions: ['js', 'json', 'jsx', 'ts', 'tsx', 'node'],
-
-	rootDir: 'source',
-
-	setupFiles: [
-		/*'jest-canvas-mock'*/
-	],
-}
