+++ packages/webapp/package.json
@@ -1,42 +1,7 @@
 {
 	"name": "webapp",
-	"main": "source/main.tsx",
-	"scripts": {
-		"build": "rollup -c",
-		"dev": "rollup -c -w",
-		"test": "jest --config jest.config.js",
-		"test:coverage": "yarn run test --collectCoverage",
-		"serve": "yarn dlx serve public",
-		"lint": "prettier --config ../../.prettierrc.json --ignore-path ../../.prettierignore --check ."
-	},
-	"dependencies": {
-		"ecsy": "^0.4.0",
-		"es": "link:./source/ES",
-		"pixi.js": "^5.3.3",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
-		"regexparam": "^1.3.0"
-	},
+	"scripts": {},
 	"devDependencies": {
-		"@babel/core": "^7.11.1",
-		"@babel/preset-env": "^7.11.0",
-		"@babel/preset-typescript": "^7.10.4",
-		"@rollup/plugin-commonjs": "^14.0.0",
-		"@rollup/plugin-node-resolve": "^8.4.0",
-		"@rollup/plugin-sucrase": "^3.1.0",
-		"@rollup/plugin-typescript": "^5.0.2",
-		"@rollup/plugin-url": "^5.0.1",
-		"@types/jest": "^26.0.9",
-		"@types/react": "^16.9.45",
-		"@types/react-dom": "^16.9.8",
-		"babel-jest": "^26.2.2",
-		"jest": "^26.2.2",
-		"prettier": "^2.0.5",
-		"rollup": "^2.23.1",
-		"rollup-plugin-livereload": "^1.3.0",
-		"rollup-plugin-serve": "^1.0.3",
-		"rollup-plugin-terser": "^6.1.0",
-		"tslib": "^2.0.1",
-		"typescript": "^3.9.7"
+		"serve": "^11.3.2"
 	}
 }
