--- packages/webapp/source/ES/pixi.ts
@@ -1,4 +1,0 @@
-import type PIXI from 'pixi.js'
-
-const pixi: typeof PIXI = window.PIXI
-export default pixi
