--- packages/webapp/source/Pages/404/index.tsx
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export default () => <h1>404 Page Not Found</h1>
