--- packages/webapp/source/Pages/Home/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import Game from '../../Components/Game'
-
-export default () => {
-	return (
-		<div id="home">
-			<h1>Home</h1>
-			<Game />
-		</div>
-	)
-}
