--- packages/webapp/tsconfig.json
@@ -1,10 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"rootDir": "source",
-		"outDir": "public/build",
-		"jsx": "react",
-		"baseUrl": "source",
-		"paths": { "es/*": ["ES/*"] }
-	}
-}
